# api/openapi/openapi.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: Topsigner
  description: |
    Modern image generator and sender for social media
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Auth
    description: Operations with authorization
  - name: Image
    description: Operations with images
  - name: Font
    description: Operations with fonts
paths:
  /api/v1/images:
    get:
      summary: Get list of user images
      description: Returns a paginated list of user images
      operationId: getImages
      tags:
        - Image
      responses:
        '200':
          description: Paginated list of images
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Upload image file
      description: Uploads a new image file
      operationId: uploadImage
      tags:
        - Image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  description: Image file to upload
                  type: string
                  format: binary  
      responses:
        '201':
          description: Image successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'  
  /api/v1/images/{id}:
    delete:
      summary: Delete image by ID
      description: Returns
      operationId: deleteImageByID
      tags:
        - Image
      responses:
        '204':
          description: Image successfully deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'  
        '500':
          $ref: '#/components/responses/InternalError'
    parameters:
      - name: id
        in: path
        required: true
        description: Image ID
        schema:
          type: string
  /api/v1/fonts:
    get:
      summary: Get list of fonts
      description: Returns a paginated list of fonts
      operationId: getFonts
      tags:
        - Font
      responses:
        '200':
          description: Paginated list of fonts
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'  
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated endpoints

  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable error description

    Image:
      type: object
      description: Image resource
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          x-order: 1
        name:
          type: string
          x-order: 2
        created_at:
          type: string
          format: date-time
          x-order: 3
        updated_at:
          type: string
          format: date-time
          x-order: 4

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    UnsupportedMediaType:
      description: Unsupported file type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'