# gen/openapi/openapi.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package openapi

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"mime/multipart"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	// Message Human-readable error description
	Message string `json:"message"`
}

// Image Image resource
type Image struct {
	ID        int64     `json:"id"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// InternalError defines model for InternalError.
type InternalError = Error

// TooManyRequests defines model for TooManyRequests.
type TooManyRequests = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// bearerAuthContextKey is the context key for BearerAuth security scheme
type bearerAuthContextKey string

// UploadImageMultipartBody defines parameters for UploadImage.
type UploadImageMultipartBody struct {
	// File Image file to upload
	File openapi_types.File `json:"file"`
}

// UploadImageMultipartRequestBody defines body for UploadImage for multipart/form-data ContentType.
type UploadImageMultipartRequestBody UploadImageMultipartBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get list of fonts
	// (GET /api/v1/fonts)
	GetFonts(w http.ResponseWriter, r *http.Request)
	// Get list of user images
	// (GET /api/v1/images)
	GetImages(w http.ResponseWriter, r *http.Request)
	// Upload image file
	// (POST /api/v1/images)
	UploadImage(w http.ResponseWriter, r *http.Request)
	// Delete image by ID
	// (DELETE /api/v1/images/{id})
	DeleteImageByID(w http.ResponseWriter, r *http.Request, id string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get list of fonts
// (GET /api/v1/fonts)
func (_ Unimplemented) GetFonts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get list of user images
// (GET /api/v1/images)
func (_ Unimplemented) GetImages(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Upload image file
// (POST /api/v1/images)
func (_ Unimplemented) UploadImage(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete image by ID
// (DELETE /api/v1/images/{id})
func (_ Unimplemented) DeleteImageByID(w http.ResponseWriter, r *http.Request, id string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetFonts operation middleware
func (siw *ServerInterfaceWrapper) GetFonts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetFonts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetImages operation middleware
func (siw *ServerInterfaceWrapper) GetImages(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetImages(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UploadImage operation middleware
func (siw *ServerInterfaceWrapper) UploadImage(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UploadImage(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteImageByID operation middleware
func (siw *ServerInterfaceWrapper) DeleteImageByID(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteImageByID(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/fonts", wrapper.GetFonts)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/images", wrapper.GetImages)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/images", wrapper.UploadImage)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/v1/images/{id}", wrapper.DeleteImageByID)
	})

	return r
}

type BadRequestJSONResponse Error

type ConflictJSONResponse Error

type InternalErrorJSONResponse Error

type TooManyRequestsResponseHeaders struct {
	RetryAfter *int
}
type TooManyRequestsJSONResponse struct {
	Body Error

	Headers TooManyRequestsResponseHeaders
}

type UnauthorizedJSONResponse Error

type GetFontsRequestObject struct {
}

type GetFontsResponseObject interface {
	VisitGetFontsResponse(w http.ResponseWriter) error
}

type GetFonts200Response struct {
}

func (response GetFonts200Response) VisitGetFontsResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type GetFonts400JSONResponse struct{ BadRequestJSONResponse }

func (response GetFonts400JSONResponse) VisitGetFontsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetFonts401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetFonts401JSONResponse) VisitGetFontsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetFonts429JSONResponse struct{ TooManyRequestsJSONResponse }

func (response GetFonts429JSONResponse) VisitGetFontsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	w.WriteHeader(429)
	_, err := buf.WriteTo(w)
	return err
}

type GetFonts500JSONResponse struct{ InternalErrorJSONResponse }

func (response GetFonts500JSONResponse) VisitGetFontsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetImagesRequestObject struct {
}

type GetImagesResponseObject interface {
	VisitGetImagesResponse(w http.ResponseWriter) error
}

type GetImages200Response struct {
}

func (response GetImages200Response) VisitGetImagesResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type GetImages400JSONResponse struct{ BadRequestJSONResponse }

func (response GetImages400JSONResponse) VisitGetImagesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetImages401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetImages401JSONResponse) VisitGetImagesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetImages429JSONResponse struct{ TooManyRequestsJSONResponse }

func (response GetImages429JSONResponse) VisitGetImagesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	w.WriteHeader(429)
	_, err := buf.WriteTo(w)
	return err
}

type GetImages500JSONResponse struct{ InternalErrorJSONResponse }

func (response GetImages500JSONResponse) VisitGetImagesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type UploadImageRequestObject struct {
	Body *multipart.Reader
}

type UploadImageResponseObject interface {
	VisitUploadImageResponse(w http.ResponseWriter) error
}

type UploadImage201JSONResponse Image

func (response UploadImage201JSONResponse) VisitUploadImageResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type UploadImage400JSONResponse struct{ BadRequestJSONResponse }

func (response UploadImage400JSONResponse) VisitUploadImageResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type UploadImage401JSONResponse struct{ UnauthorizedJSONResponse }

func (response UploadImage401JSONResponse) VisitUploadImageResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type UploadImage409JSONResponse struct{ ConflictJSONResponse }

func (response UploadImage409JSONResponse) VisitUploadImageResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type UploadImage429JSONResponse struct{ TooManyRequestsJSONResponse }

func (response UploadImage429JSONResponse) VisitUploadImageResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	w.WriteHeader(429)
	_, err := buf.WriteTo(w)
	return err
}

type UploadImage500JSONResponse struct{ InternalErrorJSONResponse }

func (response UploadImage500JSONResponse) VisitUploadImageResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteImageByIDRequestObject struct {
	ID string `json:"id"`
}

type DeleteImageByIDResponseObject interface {
	VisitDeleteImageByIDResponse(w http.ResponseWriter) error
}

type DeleteImageByID204Response struct {
}

func (response DeleteImageByID204Response) VisitDeleteImageByIDResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteImageByID400JSONResponse struct{ BadRequestJSONResponse }

func (response DeleteImageByID400JSONResponse) VisitDeleteImageByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteImageByID401JSONResponse struct{ UnauthorizedJSONResponse }

func (response DeleteImageByID401JSONResponse) VisitDeleteImageByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteImageByID429JSONResponse struct{ TooManyRequestsJSONResponse }

func (response DeleteImageByID429JSONResponse) VisitDeleteImageByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	w.WriteHeader(429)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteImageByID500JSONResponse struct{ InternalErrorJSONResponse }

func (response DeleteImageByID500JSONResponse) VisitDeleteImageByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get list of fonts
	// (GET /api/v1/fonts)
	GetFonts(ctx context.Context, request GetFontsRequestObject) (GetFontsResponseObject, error)
	// Get list of user images
	// (GET /api/v1/images)
	GetImages(ctx context.Context, request GetImagesRequestObject) (GetImagesResponseObject, error)
	// Upload image file
	// (POST /api/v1/images)
	UploadImage(ctx context.Context, request UploadImageRequestObject) (UploadImageResponseObject, error)
	// Delete image by ID
	// (DELETE /api/v1/images/{id})
	DeleteImageByID(ctx context.Context, request DeleteImageByIDRequestObject) (DeleteImageByIDResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetFonts operation middleware
func (sh *strictHandler) GetFonts(w http.ResponseWriter, r *http.Request) {
	var request GetFontsRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetFonts(ctx, request.(GetFontsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetFonts")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetFontsResponseObject); ok {
		if err := validResponse.VisitGetFontsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetImages operation middleware
func (sh *strictHandler) GetImages(w http.ResponseWriter, r *http.Request) {
	var request GetImagesRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetImages(ctx, request.(GetImagesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetImages")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetImagesResponseObject); ok {
		if err := validResponse.VisitGetImagesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// UploadImage operation middleware
func (sh *strictHandler) UploadImage(w http.ResponseWriter, r *http.Request) {
	var request UploadImageRequestObject

	if reader, err := r.MultipartReader(); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode multipart body: %w", err))
		return
	} else {
		request.Body = reader
	}

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.UploadImage(ctx, request.(UploadImageRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UploadImage")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(UploadImageResponseObject); ok {
		if err := validResponse.VisitUploadImageResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// DeleteImageByID operation middleware
func (sh *strictHandler) DeleteImageByID(w http.ResponseWriter, r *http.Request, id string) {
	var request DeleteImageByIDRequestObject

	request.ID = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteImageByID(ctx, request.(DeleteImageByIDRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteImageByID")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteImageByIDResponseObject); ok {
		if err := validResponse.VisitDeleteImageByIDResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}